backend/requirements.txt
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0
transformers>=4.57.0
huggingface-hub>=0.36.0
accelerate>=1.11.0
fastapi>=0.121.0
uvicorn>=0.38.0
python-multipart>=0.0.20
peft>=0.10.0
trl>=0.9.0
datasets>=2.20.0
llama-cpp-python
crewai
pyautogui
selenium
webdriver-manager
networkx

